Refactor table stats into separate component
